backend/requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
Pillow>=10.0.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=13.0
